always insert tracking code
previously we'd not insert the tracking code when there was no trace id.
However, we want to always insert the tracking code but only exclude the
ineum call for trace ids when it doesnt exist

also updates the screenshot because it contained a token. not relevant
for security, but still shouldn't be there.